Refactor notes endpoints using APIRouter
